(editor) Extract deepMerge and shared styles to utils
